# specs/main/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Embodied AI Systems RAG API
  description: API for the Embodied AI Systems Book RAG chatbot
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-12-10T10:00:00Z"

  /chat:
    post:
      summary: Main chat endpoint for RAG responses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: User query about book content
                  example: "What are ROS 2 nodes?"
                  minLength: 1
                  maxLength: 1000
                use_selected_text:
                  type: boolean
                  description: Whether to restrict answers to selected text only
                  example: false
                selected_text:
                  type: string
                  description: Text selected by user for context (if use_selected_text is true)
                  example: "A node is a process that performs computation..."
                  nullable: true
              required:
                - query
      responses:
        '200':
          description: Chat response with citations
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: AI-generated response
                    example: "Based on the book content: ROS 2 nodes are processes that perform computation..."
                  citations:
                    type: array
                    items:
                      type: string
                    description: Citations for the response
                    example: ["Module: Module 1: Robotic Nervous System (ROS 2), Chapter: ROS 2 Basics"]
                  query:
                    type: string
                    description: Original query
                    example: "What are ROS 2 nodes?"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-12-10T10:00:00Z"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /retrieve:
    post:
      summary: Semantic search in the knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Search query
                  example: "URDF basics"
                  minLength: 1
                limit:
                  type: integer
                  description: Number of results to return
                  example: 5
                  default: 5
                  minimum: 1
                  maximum: 20
              required:
                - query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "uuid-string"
                        text:
                          type: string
                          example: "URDF (Unified Robot Description Format) is an XML format..."
                        module:
                          type: string
                          example: "Module 1: Robotic Nervous System (ROS 2)"
                        chapter:
                          type: string
                          example: "ROS 2 Basics"
                        score:
                          type: number
                          format: float
                          example: 0.95
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /embed:
    post:
      summary: Generate embeddings for text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: Text to generate embeddings for
                  example: "This is a sample text for embedding"
              required:
                - text
      responses:
        '200':
          description: Embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  embeddings:
                    type: array
                    items:
                      type: number
                      format: float
                    description: Embedding vector
                    example: [0.1, 0.2, 0.3, ...]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /add_document:
    post:
      summary: Add a document to the knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Document added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Document added successfully"
                  qdrant_id:
                    type: string
                    example: "uuid-string"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /:
    get:
      summary: Root endpoint
      responses:
        '200':
          description: API status
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Embodied AI Systems RAG API"
                  status:
                    type: string
                    example: "running"

components:
  schemas:
    Document:
      type: object
      properties:
        id:
          type: string
          description: Document identifier
          example: "doc-001"
        module:
          type: string
          description: Module identifier
          example: "Module 1: Robotic Nervous System (ROS 2)"
        chapter:
          type: string
          description: Chapter identifier
          example: "ROS 2 Basics"
        anchor:
          type: string
          description: Section anchor (optional)
          example: "nodes-and-topics"
          nullable: true
        text:
          type: string
          description: Document text content
          example: "This is the content of the document..."
        qdrant_id:
          type: string
          description: Qdrant vector ID (optional)
          example: "uuid-string"
          nullable: true
      required:
        - id
        - module
        - chapter
        - text